+++ F1H43C_EEJYN9/Battleship.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "F1H43C_EEJYN9", "Battleship\F1H43C_EEJYN9.csproj", "{5112EF0D-FB04-4D62-A58E-8004D60178C0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Battleship_Tests", "Battleship_Tests\Battleship_Tests.csproj", "{E6188C9E-0743-4BDB-80F3-BAC715CA12A6}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{5112EF0D-FB04-4D62-A58E-8004D60178C0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5112EF0D-FB04-4D62-A58E-8004D60178C0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5112EF0D-FB04-4D62-A58E-8004D60178C0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5112EF0D-FB04-4D62-A58E-8004D60178C0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E6188C9E-0743-4BDB-80F3-BAC715CA12A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E6188C9E-0743-4BDB-80F3-BAC715CA12A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E6188C9E-0743-4BDB-80F3-BAC715CA12A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E6188C9E-0743-4BDB-80F3-BAC715CA12A6}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
