+++ F1H43C_EEJYN9/Battleship/Core/PreferencesManager.cs
