+++ F1H43C_EEJYN9/Battleship/Entities/GameData.cs
@@ -1,0 +1,13 @@
+namespace F1H43C_EEJYN9.Entities;
+
+public record GameData
+{
+    public  DateTime Date { get; init; }
+    public string Winner { get; init; }
+
+    public GameData(DateTime date, string winner)
+    {
+        Date = date;
+        Winner = winner;
+    }
+}
