+++ F1H43C_EEJYN9/Battleship/Entities/Interfaces/IPreferenceManager.cs
